# Slight increase in testing speed