We might want to consider testing incrementally