@obriente 2020-11-05
LGTM, the changes to fermion_partitioning are fine.